refactor: Move binary installation to it's own crate
