+++ Cargo.toml
@@ -9,6 +9,9 @@ readme = "README.md"
 categories = ["wasm"]
 documentation = "https://rustwasm.github.io/wasm-pack/"
 
+[workspace]
+members = [".", "binary-install"]
+
 [dependencies]
 atty = "0.2.11"
 cargo_metadata = "0.6.0"
@@ -32,6 +35,7 @@ tar = "0.4.16"
 toml = "0.4"
 which = "2.0.0"
 zip = "0.4.2"
+binary-install = { version = "0.1.0", path = "./binary-install" }
 
 [dev-dependencies]
 tempfile = "3"
