+++ binary-install/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "binary-install"
+version = "0.1.0"
+authors = ["Jesper Håkansson <jesper@jesperh.se>"]
+
+[dependencies]
+curl = "0.4.13"
+failure = "0.1.2"
+flate2 = "1.0.2"
+slog = "2.3"
+tar = "0.4.16"
+which = "2.0.0"
+zip = "0.4.2"
+
