+++ tests/all/main.rs
@@ -4,6 +4,7 @@ extern crate serde_derive;
 extern crate serde_json;
 #[macro_use]
 extern crate slog;
+extern crate binary_install;
 extern crate structopt;
 extern crate tempfile;
 extern crate wasm_pack;
